ref(users): move users related web resources to users directory
Moves user related web resources and adds typing where needed.

ref(issue # here)
