+++ src/sentry/users/web/user_avatar.py
