+++ src/sentry/web/urls.py
@@ -13,9 +13,11 @@ from sentry.auth.providers.saml2.provider import SAML2AcceptACSView, SAML2Metada
 from sentry.charts.endpoints import serve_chartcuterie_config
 from sentry.integrations.web.doc_integration_avatar import DocIntegrationAvatarPhotoView
 from sentry.integrations.web.organization_integration_setup import OrganizationIntegrationSetupView
+from sentry.users.web import accounts
+from sentry.users.web.account_identity import AccountIdentityAssociateView
+from sentry.users.web.user_avatar import UserAvatarPhotoView
 from sentry.web import api
-from sentry.web.frontend import accounts, generic
-from sentry.web.frontend.account_identity import AccountIdentityAssociateView
+from sentry.web.frontend import generic
 from sentry.web.frontend.auth_channel_login import AuthChannelLoginView
 from sentry.web.frontend.auth_close import AuthCloseView
 from sentry.web.frontend.auth_login import AuthLoginView
@@ -47,7 +49,6 @@ from sentry.web.frontend.setup_wizard import SetupWizardView
 from sentry.web.frontend.shared_group_details import SharedGroupDetailsView
 from sentry.web.frontend.sudo import SudoView
 from sentry.web.frontend.twofactor import TwoFactorAuthView, u2f_appid
-from sentry.web.frontend.user_avatar import UserAvatarPhotoView
 
 __all__ = ("urlpatterns",)
