+++ tests/sentry/users/web/test_user_avatar.py
